fix(simulation,reference-model): four Codex-connector findings on N12, all real
1. `b_min` ceiled a float, not the Decimal quotient, so above float's 53-bit
   mantissa it rounded DOWN: at stake 10,000,005,521 it returned ...317 where
   the exact ceiling is ...318. R-7 puts rounding against the party relying on
   the claim, so a short b_min underfunds the certification it certifies.

2. The competing-venue control ran at the CALIBRATED flow cap (16) while the
   primary population it is compared against ran at the probe cap (17), so
   `control_matches_primary` compared two populations differing in a second
   variable. Same class of confound as the partial-epoch-slate sampling fixed
   earlier in this branch; same fix, hold everything but the diversion fixed.

3. `--check` validated `liveness_loss_at_arming` by truthiness alone. It is
   the only MEASURED half of the Phase-4 denial analysis — the security half
   is a screen — so it is now derived from the rungs and compared field by
   field, including the `liveness` detail block.

4. The artifact's own `assumptions.competing_venue` string still said the rung
   was "derived" and capped diversion at one half, contradicting the amended
   16 §8.4 and the adjacent `flow_model`. That is a false claim baked into
   published evidence, which is worse than one in prose.

Artifact regeneration follows; the committed artifact is stale against this
source until it lands.

+++ reference-model/src/bleavit_reference_model/service_economics.py
@@ -119,7 +119,14 @@ def b_min(stake: Decimal, epsilon: Decimal) -> Decimal:
     `SECURITY_FACTOR = 3`. Rounds UP, against the party relying on the claim.
     """
     denominator = 2 * _ln(Decimal("0.5") / (Decimal("0.5") - epsilon))
-    return Decimal(math.ceil(float(3 * stake / denominator)))
+    # Ceil the Decimal quotient directly. Routing it through `float` first --
+    # as an earlier revision did -- silently rounds DOWN once the quotient
+    # exceeds float's 53-bit mantissa: at stake 10,000,005,521 and eps 0.05 it
+    # returns ...317 where the exact ceiling is ...318. One unit is trivial in
+    # absolute terms and the direction is not: R-7 puts rounding against the
+    # party relying on the claim, and here that is the client's counterparty,
+    # so a short b_min underfunds the certification it certifies.
+    return Decimal(math.ceil(3 * stake / denominator))
 
 
 def b_min_superseded(stake: Decimal, epsilon: Decimal) -> Decimal:
@@ -132,7 +139,7 @@ def b_min_superseded(stake: Decimal, epsilon: Decimal) -> Decimal:
     denominator = 2 * _ln(
         (Decimal("0.5") + epsilon) / (Decimal("0.5") - epsilon)
     )
-    return Decimal(math.ceil(float(3 * stake / denominator)))
+    return Decimal(math.ceil(3 * stake / denominator))
 
 
 def client_subsidy(stake: Decimal, epsilon: Decimal) -> Decimal:
